+++ lib/pio/open_flow13.rb
@@ -13,7 +13,18 @@ require 'pio/open_flow13/packet_out'
 require 'pio/open_flow13/stats_request'
 
 # Actions
+require 'pio/open_flow/nicira_resubmit'
+require 'pio/open_flow/nicira_resubmit_table'
+require 'pio/open_flow13/nicira_reg_load'
+require 'pio/open_flow13/nicira_reg_move'
+require 'pio/open_flow13/nicira_send_out_port'
 require 'pio/open_flow13/send_out_port'
+require 'pio/open_flow13/set_arp_operation'
+require 'pio/open_flow13/set_arp_sender_hardware_address'
+require 'pio/open_flow13/set_arp_sender_protocol_address'
+require 'pio/open_flow13/set_destination_mac_address'
+require 'pio/open_flow13/set_metadata'
+require 'pio/open_flow13/set_source_mac_address'
 
 # Instructions
 require 'pio/open_flow13/apply'