+++ lib/pio/open_flow13/goto_table.rb
@@ -28,6 +28,7 @@ module Pio
     def_delegators :@format, :instruction_length
     def_delegators :@format, :table_id
     def_delegators :@format, :to_binary_s
+    def_delegator :@format, :to_binary_s, :to_binary
 
     def initialize(table_id)
       @format = Format.new(table_id: table_id)