+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 ## develop (unreleased)
 
+## 0.30.0 (11/17/2015)
+### New features
+* [#281](https://github.com/trema/pio/pull/281): Add `NiciraRegLoad` action.
+* [#282](https://github.com/trema/pio/pull/282): Add `NiciraResubmit` action.
+* [#283](https://github.com/trema/pio/pull/283): Add `SetMetadata` action.
+* [#284](https://github.com/trema/pio/pull/284): Enable masking with reg0-reg7.
+* [#285](https://github.com/trema/pio/pull/285): Add Nicira `SendOutPort` action.
+
+
+## 0.29.0 (11/11/2015)
+### New features
+* [#274](https://github.com/trema/pio/pull/274): Add `SetSourceMacAddress` action.
+* [#275](https://github.com/trema/pio/pull/275): Add `SetDestinationMacAddress` action.
+* [#276](https://github.com/trema/pio/pull/276): Add `SetArpOperation` action.
+* [#277](https://github.com/trema/pio/pull/277): Add `SetArpSenderProtocolAddress` action.
+* [#278](https://github.com/trema/pio/pull/278): Add `SetArpSenderHardwareAddress` action.
+* [#280](https://github.com/trema/pio/pull/280): Add `NiciraRegMove` action.
+
 
 ## 0.28.1 (11/5/2015)
 ### Changes